Created SFP directory
